Simple executor now is FIFO executor
